[erros] Move example to non _test package #107
- just a workaround to let travis pass for now
